For #42779: Uses a “full” caching policy when bootstrapping.
This will cause all engines and their respective bundles to be downloaded if they’re not already on disk. This is needed because the desktop integration provided by desktop server queries all launcher commands from launch app, which requires all engines for the given environment to be on disk, as the scanning logic used to discover DCCs is housed in the engines.