For #43057: Removes synchronous execution of get_actions RPC commands.
Given that the sqlite cache file is stored on a local disk, the issue of concurrent writing connections to the database is a non issue. Because of this, we should be able to allow get_actions calls (which might call cache_actions) to run in parallel. We know that this should work in a stabile way, because it’s how the path cache works in tk-core.