For #43236: Engine init/bootstrap errors handled gracefully.
Rather than stopping the get_actions process and raising an error to the client when a bootstrap fails for one config, we continue onto other configs and reply with a list of actions for those configs that bootstrap properly. We log the information necessary to debug any bootstrap issues for the problematic configs.